voice_app/requirements.txt
fastapi
uvicorn
faster-whisper
sentence-transformers
faiss-cpu
pypdf
soundfile
pyttsx3
ollama

